Updated Activity scripting for ACraft carriable mass
As MaxMass has been changed to MaxInventoryMass and now no longer includes the mass of the Craft itself, some lua calculations needed to be adjusted to compensate. Mostly includes any script that orders a loadout in a random craft and abides by mass constraints. Some maths is approximate, such as the loop that adds inventory items only stopping *after* the max mass has been exceeded (this behaviour is unchanged, remains from legacy code). Sugested refactor later.
As `MaxInventoryMass == 0`now indicates a Craft with o cargo space (rather than undefined cargo space) when a random craft is selected with `MaxInventoryMass == 0`, it is disarded and the drop defalts to a craft from Base.rte.